Refactoring PortScanner into one class
